fix(variants): decimal pack sizes (2.25 ltr) were rejected, and truncated on read
dto.ProductVariantUnit.Value was int while domain.ProductVariant.VolumeValue
is float64 over a numeric column. Two bugs from one type:

  write: {"value":2.25} -> 400 "cannot unmarshal number 2.25 into Go struct
         field ProductVariantUnit.volume.value of type int"
  read:  a 1.5 ltr row in the DB came back over the API as 1

The read half is the worse of the two -- it answered wrong instead of
erroring, so nothing surfaced it.

Everything else already handled decimals: packLabel formats with
FormatFloat, UnitPricePaise takes float64 and divides by volumeValue*1000
for litres, the column is numeric. Only the DTO narrowed, and the two
float64()/int() casts existed solely to bridge that mistake -- they go with
it.

2.25 ltr is a real Coca-Cola pack size on blinkit (item 18314). Whole
numbers still serialise as 2, not 2.0.

Verified live: 2.25 ltr creates, the 1.5 ltr row now reads back as 1.5.

+++ internal/dto/product_variant_item.go
@@ -5,8 +5,10 @@ package dto
 // @Property unit: Unit of measurement
 // @Description Volume/weight of a product variant, e.g. "500 gm".
 type ProductVariantUnit struct {
-	// Numeric quantity of the unit
-	Value int `json:"value"`
+	// Numeric quantity of the unit. Float, not int: 2.25 ltr and 1.5 ltr are real
+	// pack sizes, and domain.ProductVariant.VolumeValue is float64 over a numeric
+	// column. Whole numbers still serialise as 2, not 2.0.
+	Value float64 `json:"value"`
 	// Unit of measurement
 	Unit string `json:"unit" enums:"kg,gm,ltr,ml,pcs"`
 }
